Fixed merged two branches

+++ src/controller/states/Walking.h
@@ -14,7 +14,7 @@ namespace states {
     class WalkingState : public State {
     public:
         void update_position(int type, FSM *fsm, int direction) {
-            if (Keyboard::getInstance().get_key() == (type)) {
+            if ((int)Keyboard::getInstance().get_key() == (type)) {
                 FSMGame *fsm1 = dynamic_cast<FSMGame *>(fsm);
                 auto current_room_real_pointer = fsm1->this_room.get();
                 if (typeid(*current_room_real_pointer) == typeid(Room)) {
